proyecto actualizado

+++ src/main/java/services/MessageService.java
@@ -3,6 +3,7 @@ package services;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 
@@ -34,9 +35,12 @@ public class MessageService {
 
 	//Simple CRUD Methods
 	public Message create() {
+		//Assert.isTrue
 		Message result;
 		result = new Message();
 		final Actor sender = this.actorService.getActorLogged();
+		final Date actualDate = Calendar.getInstance().getTime();
+		result.setSendDate(actualDate);
 		result.setSender(sender);
 		result.setRecipient(new ArrayList<Actor>());
 		return result;
@@ -62,12 +66,13 @@ public class MessageService {
 		final List<Actor> recipients = (List<Actor>) result.getRecipient();
 		for (int i = 0; i < recipients.size(); i++) {
 			final Actor a = recipients.get(i);
-			Assert.isTrue(a.getId() == 0);
+			Assert.isTrue(a.getId() != 0);
 		}
 		//Set sender message
 		result.setSender(sender);
 		//meter message outbox sender
-		final Box outBoxS = (Box) sender.getBoxes().toArray()[1];
+		final List<Box> boxesS = (List<Box>) sender.getBoxes();
+		final Box outBoxS = boxesS.get(1);
 		final List<Message> m = (List<Message>) outBoxS.getMessages();
 		m.add(result);
 		outBoxS.setMessages(m);
@@ -84,20 +89,22 @@ public class MessageService {
 		if (msgIsSpam)
 			for (int i = 0; i < recipients.size(); i++) {
 				final Actor a = recipients.get(i);
-				final Box spamBoxRx = (Box) a.getBoxes().toArray()[3];
+				final List<Box> boxesR = (List<Box>) a.getBoxes();
+				final Box spamBoxRx = boxesR.get(3);
 				final List<Message> m1 = (List<Message>) spamBoxRx.getMessages();
 				m1.add(result);
 				spamBoxRx.setMessages(m1);
-				this.boxService.save(spamBoxRx);
+				this.boxService.saveRecipient(spamBoxRx, a);
 			}
 		else
 			for (int i = 0; i < recipients.size(); i++) {
 				final Actor a = recipients.get(i);
-				final Box inBoxRx = (Box) a.getBoxes().toArray()[3];
+				final List<Box> boxesR = (List<Box>) a.getBoxes();
+				final Box inBoxRx = boxesR.get(0);
 				final List<Message> m1 = (List<Message>) inBoxRx.getMessages();
 				m1.add(result);
 				inBoxRx.setMessages(m1);
-				this.boxService.save(inBoxRx);
+				this.boxService.saveRecipient(inBoxRx, a);
 			}
 		//Guardar mensaje en BD
 		return result;
@@ -105,8 +112,17 @@ public class MessageService {
 
 	public void delete(final Message message) {
 		Assert.notNull(message);
-		Assert.isTrue(message.getId() == 0);
-		this.messageRepository.delete(message);
+		Assert.isTrue(message.getId() != 0);
+		final List<Box> boxes = new ArrayList<Box>();
+		boxes.addAll(this.boxService.findAll());
+		boolean msgInAnyBox = false;
+		for (int i = 0; i < boxes.size(); i++)
+			if (boxes.get(i).getMessages().contains(message)) {
+				msgInAnyBox = true;
+				break;
+			}
+		if (msgInAnyBox)
+			this.messageRepository.delete(message);
 	}
 
 	//Other business methods
@@ -114,7 +130,7 @@ public class MessageService {
 	public void moveMessage(final Message message, final Box boxR) {
 		//Asserts e inicializaciones
 		Assert.notNull(message);
-		Assert.isTrue(message.getId() == 0);
+		Assert.isTrue(message.getId() != 0);
 		final Actor a = message.getSender();
 		final List<Box> boxesActor = (List<Box>) a.getBoxes();
 		Box originBox = null;
@@ -128,7 +144,6 @@ public class MessageService {
 			if (boxesActor.get(i).equals(boxR))
 				boxInActor = true;
 		}
-		Assert.isTrue(!msgInActor || !boxInActor);
 		//Mover mensage
 		final List<Message> oM = (List<Message>) originBox.getMessages();
 		oM.remove(message);
@@ -144,7 +159,7 @@ public class MessageService {
 	public void deleteMessage(final Message message) {
 		//Asserts e inicializaciones
 		Assert.notNull(message);
-		Assert.isTrue(message.getId() == 0);
+		Assert.isTrue(message.getId() != 0);
 		final Actor a = message.getSender();
 		final List<Box> boxesActor = (List<Box>) a.getBoxes();
 		Box originBox = null;
@@ -154,7 +169,7 @@ public class MessageService {
 				originBox = boxesActor.get(i);
 				msgInActor = true;
 			}
-		Assert.isTrue(!msgInActor);
+		Assert.isTrue(msgInActor);
 		//Si no es trashbox, mover a trashbox
 		if (boxesActor.get(2) != originBox) {
 			final Box trashBox = boxesActor.get(2);
@@ -168,28 +183,36 @@ public class MessageService {
 			this.boxService.save(trashBox);
 			//Si trashbox, borrar de actor
 		} else
-			for (int i = 0; i < a.getBoxes().size(); i++)
-				if (boxesActor.get(i).getMessages().contains(message)) {
-					final Box boxWithMsg = boxesActor.get(i);
-					final List<Message> msgs = (List<Message>) boxWithMsg.getMessages();
-					msgs.remove(message);
-					boxWithMsg.setMessages(msgs);
-					this.boxService.save(boxWithMsg);
-					/*
-					 * final Box boxWithMsg = boxesActor.get(i);
-					 * boxesActor.remove(boxWithMsg);
-					 * boxWithMsg.getMessages().remove(message);
-					 * boxesActor.add(boxWithMsg);
-					 * a.setBoxes(boxesActor);
-					 * this.boxService.save(boxWithMsg);
-					 */
+			/*
+			 * for (int i = 0; i < a.getBoxes().size(); i++)
+			 * if (boxesActor.get(i).getMessages().contains(message)) {
+			 * final Box boxWithMsg = boxesActor.get(i);
+			 * final List<Message> msgs = (List<Message>) boxWithMsg.getMessages();
+			 * msgs.remove(message);
+			 * boxWithMsg.setMessages(msgs);
+			 * this.boxService.save(boxWithMsg);
+			 * /*
+			 * final Box boxWithMsg = boxesActor.get(i);
+			 * boxesActor.remove(boxWithMsg);
+			 * boxWithMsg.getMessages().remove(message);
+			 * boxesActor.add(boxWithMsg);
+			 * a.setBoxes(boxesActor);
+			 * this.boxService.save(boxWithMsg);
+			 */
+			for (final Box b : a.getBoxes())
+				if (b.getMessages().contains(message)) {
+					final Collection<Message> messages = b.getMessages();
+					messages.remove(message);
+					b.setMessages(messages);
+					this.boxService.save(b);
+					this.delete(message);
 				}
 	}
 
 	public Message adminMessage(final Message message) {
 		//Asserts e inicializaciones
 		Assert.notNull(message);
-		Assert.isTrue(message.getId() != 0);
+		Assert.isTrue(message.getId() == 0);
 		//Mandar mensaje a todos los actores del sistema
 		final Actor admin = this.administratorService.findOne(this.actorService.getActorLogged().getId());
 		final List<Actor> recipient = new ArrayList<Actor>();
